+++ src/main/kotlin/nl/deschepers/laraveltinker/settings/PluginSettingsConfigurable.kt
@@ -7,7 +7,7 @@ import javax.swing.JComponent
 class PluginSettingsConfigurable : Configurable {
     private var pluginSettingsWindow: PluginSettingsWindow? = null
 
-    override fun createComponent(): JComponent? {
+    override fun createComponent(): JComponent {
         pluginSettingsWindow = PluginSettingsWindow()
         return pluginSettingsWindow!!.getPanel()
     }
@@ -15,19 +15,22 @@ class PluginSettingsConfigurable : Configurable {
     override fun isModified(): Boolean {
         val settings = PluginSettings.getInstance()
         return pluginSettingsWindow!!.showExecutionStarted != settings.showExecutionStarted ||
-            pluginSettingsWindow!!.showExecutionEnded != settings.showExecutionEnded
+            pluginSettingsWindow!!.showExecutionEnded != settings.showExecutionEnded ||
+            pluginSettingsWindow!!.useWordWrapping != settings.useWordWrapping
     }
 
     override fun apply() {
         val settings = PluginSettings.getInstance()
         settings.showExecutionStarted = pluginSettingsWindow!!.showExecutionStarted
         settings.showExecutionEnded = pluginSettingsWindow!!.showExecutionEnded
+        settings.useWordWrapping = pluginSettingsWindow!!.useWordWrapping
     }
 
     override fun reset() {
         val settings = PluginSettings.getInstance()
         pluginSettingsWindow!!.showExecutionStarted = settings.showExecutionStarted
         pluginSettingsWindow!!.showExecutionEnded = settings.showExecutionEnded
+        pluginSettingsWindow!!.useWordWrapping = settings.useWordWrapping
     }
 
     override fun getDisplayName(): String {